pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pips-solver"
version = "1.0.0"
description = "Per-Instance Program Synthesis (PIPS) - An adaptive reasoning baseline that alternates between chain-of-thought and executable program synthesis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PIPS Development Team", email = "contact@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["llm", "code-generation", "ai", "problem-solving", "iterative"]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-genai>=0.2.0",
    "flask>=2.0.0",
    "flask-socketio>=5.0.0",
    "pillow>=8.0.0",
    "timeout-decorator>=0.5.0",
    "python-socketio[client]>=5.0.0",
]

[project.optional-dependencies]
web = [
    "flask>=2.0.0",
    "flask-socketio>=5.0.0",
    "python-socketio[client]>=5.0.0",
]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.0.0",
]
all = [
    "pips-solver[web,dev]"
]

[project.scripts]
pips = "pips.__main__:main"

[project.urls]
Homepage = "https://github.com/example/pips"
Repository = "https://github.com/example/pips"
Issues = "https://github.com/example/pips/issues"
Documentation = "https://github.com/example/pips#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/pips"]

[tool.hatch.build.targets.sdist]
include = [
    "/src/pips",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
